requirements.txt
pymupdf
fpdf2
requests
ebooklib
lxml
